Further results with same code prove possible convergence with sufficient nodes
After testing for convergent solutions, it was found that for a square matrix (N=M), a 14 by 14 matrix was the smallest possible matrix that would produce a convergent solution. Anything 13 by 13 or smaller produced divergent solutions as previously mentioned. This edit also includes instructions on what to do following this edit/commit.