Surface and Contour plots
Included surface and contour plots for solution visualization and verified that, if approximation were done for a Poisson's equation, solution would converge at all values. I.e.: Took lambda = 0 and ran, proving values did not diverge with small grid sizes. The divergence, therefore, is a consequence of the Helmholtz equation.